rename & fix type annotation
